Move ViewStateFrame to core-android and use it for modal view state.
