+++ workflow-ui/backstack-android/api/backstack-android.api
@@ -64,25 +64,6 @@ public final class com/squareup/workflow1/ui/backstack/ViewStateCache$SavedState
 	public synthetic fun newArray (I)[Ljava/lang/Object;
 }
 
-public final class com/squareup/workflow1/ui/backstack/ViewStateFrame : android/os/Parcelable {
-	public static final field CREATOR Lcom/squareup/workflow1/ui/backstack/ViewStateFrame$CREATOR;
-	public fun <init> (Ljava/lang/String;)V
-	public synthetic fun <init> (Ljava/lang/String;Landroid/util/SparseArray;Landroid/os/Bundle;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public fun describeContents ()I
-	public final fun destroyOnDetach ()V
-	public final fun performRestore (Landroid/view/View;)V
-	public final fun performSave (Landroid/view/View;)V
-	public static synthetic fun performSave$default (Lcom/squareup/workflow1/ui/backstack/ViewStateFrame;Landroid/view/View;ILjava/lang/Object;)V
-	public fun writeToParcel (Landroid/os/Parcel;I)V
-}
-
-public final class com/squareup/workflow1/ui/backstack/ViewStateFrame$CREATOR : android/os/Parcelable$Creator {
-	public fun createFromParcel (Landroid/os/Parcel;)Lcom/squareup/workflow1/ui/backstack/ViewStateFrame;
-	public synthetic fun createFromParcel (Landroid/os/Parcel;)Ljava/lang/Object;
-	public fun newArray (I)[Lcom/squareup/workflow1/ui/backstack/ViewStateFrame;
-	public synthetic fun newArray (I)[Ljava/lang/Object;
-}
-
 public final class com/squareup/workflow1/ui/backstack/databinding/ViewStackLayoutBinding : androidx/viewbinding/ViewBinding {
 	public final field viewStack Lcom/squareup/workflow1/ui/backstack/BackStackContainer;
 	public static fun bind (Landroid/view/View;)Lcom/squareup/workflow1/ui/backstack/databinding/ViewStackLayoutBinding;
